Pipeline cleanup: foundations refactor, in-house IO, memory audit, dead-code purge
A grab-bag commit covering several themes that landed together on the
mitsuba-parity branch. Each block below is independently reviewable.

Foundations:
- Replace happly + stb_image with miniply (~10× faster PLY parsing)
- In-house RGBE/.hdr parser decoding straight into pinned memory
  (eliminates the pageable→pinned memcpy and stb's global flip flag)
- Modularize src/thesis/host/utils/io.cpp into io/{file,exr,hdr,ply}.cpp
- Namespace prim_idx_t under thesis::common::utils (kept uint16_t cap)
- Replace #ifdef DEVICE with __CUDA_ARCH__; drop Quake III rsqrt host
  fallback in favor of 1/std::sqrt; remove duplicate THESIS_HOSTDEVICE
- static_assert MAX_PRIMITIVES <= max(prim_idx_t)

Memory audit:
- M1: synchronize EnvMap stream before pinned host dies (latent UAF)
- M2: keep AS out_ alive across optixAccelCompact via std::move +
  destructor-ordered cudaFreeAsync (latent UAF)
- M3: gate Welford variance buffer on ENABLE_ADAPTIVE_SAMPLING (saves
  16 MB at 4K + per-sample writes when adaptive sampling is off)
- M4: gate AOV buffers on runtime denoise flag (saves 2×W×H×16B + per-
  sample writes when not denoising)
- M5: drop per-batch synchronize (~9% wall-clock improvement; batches
  now queue back-to-back on the Main stream)
- M9: drop redundant synchronize after optixDenoiserInvoke (next op on
  same stream serializes naturally)
- M11: HostHint::WriteCombined for upload-only buffers (primitives,
  instances, camera_active_prims, env CDFs); routes to
  cudaHostAllocWriteCombined for ~2× faster host→device staging
- M12: convert single-element OnBoth buffers (compacted_size,
  launch_params) to OnDeviceOnly + plain host mirror; eliminates the
  4 KB pinned-page tax for byte-sized objects

Code quality:
- Delete Stream::addCallback (dead code with two latent bugs: leak on
  cudaStreamAddCallback failure, gratuitous std::function indirection)
- Delete Context::Guard move ops (double-pop bug — destructor pops
  unconditionally; Guard is only used as scope-bound RAII anyway)
- Slim pch.h: remove CLI11 (used only by config), tinyexr (used only
  by exr.cpp), iostream + unordered_map (zero users); split
  Config::parse body into src/thesis/host/app/config.cpp
- static_assert sizeof(Mat3x4) == sizeof(OptixInstance::transform) so
  future drift breaks at compile time, not at runtime

Cleanup:
- Remove unused device functions integrate_primitives,
  compute_optical_depth_along_ray (sampling.cuh) and Primitive::
  optical_depth(ray, t0) (primitive.h)
- Remove stale plans/, docs/{TODO,denoiser,hybrid-wavefront-plan,jorge,
  nee,quality-roadmap}.md, jorge_python.py, .claude_context_summary.md

+++ include/thesis/device/params/image.h
@@ -13,10 +13,10 @@ namespace params {
 
 // Device-side POD struct for image buffer (no RAII, same size on host and device)
 struct THESIS_ALIGNMENT Image {
-    float4* variance_ = nullptr;  // Running M2 (sum of squared deviations) for Welford's algorithm
+    float4* variance_ = nullptr;  // Welford M2; null when adaptive sampling is disabled (skip writes)
     float4* mean_ = nullptr;      // Running mean for Welford's algorithm
-    size_t* sample_counts_ =
-        nullptr;  // Number of samples taken per pixel (size_t for buffer compatibility)
+    uint16_t* sample_counts_ =
+        nullptr;  // Per-pixel sample count (max 65535 spp; widen if you ever need more)
 
     // Auxiliary outputs ("AOVs") for the OptiX denoiser as guide layers.
     // Both are running per-pixel means across SPP (no Welford M2 needed — we just need
